API for Unblock user

+++ backend/Controllers/User/user.js
@@ -421,7 +421,7 @@ exports.blockedList = (req, res) => {
               });
             }
 
-            UserFollowers.deleteOne({ follower_id: blockUserId }, (err) => {
+            UserFollowers.deleteOne({ follower_id: blockUserId, userId: userId }, (err) => {
               if (err) {
                 return res.json({
                   data: null,
@@ -431,7 +431,7 @@ exports.blockedList = (req, res) => {
               }
             });
 
-            UserFollowing.deleteOne({ following_id: blockUserId }, (err) => {
+            UserFollowing.deleteOne({ following_id: blockUserId, userId: userId }, (err) => {
               if (err) {
                 return res.json({
                   data: null,
@@ -452,6 +452,104 @@ exports.blockedList = (req, res) => {
     }
   );
 };
+
+exports.unblockUser = (req, res) => {
+  userId = req.profile._id;
+  unblockUserId = req.params.unblockUserId;
+
+  if (!unblockUserId) {
+    return res.json({
+      data: null,
+      status: "error",
+      error: "Unblock user id is required.",
+    });
+  }
+
+  if (!userId) {
+    return res.json({
+      data: null,
+      status: "error",
+      error: "UserId is required.",
+    });
+  }
+
+  if (!Helper.isMongoId(userId)) {
+    return res.json({
+      data: null,
+      status: "error",
+      error: "UserId is invalid.",
+    });
+  }
+
+  if (!Helper.isMongoId(unblockUserId)) {
+    return res.json({
+      data: null,
+      status: "error",
+      error: "Unblock user id is invalid.",
+    });
+  }
+
+  if (userId === unblockUserId) {
+    return res.json({
+      data: null,
+      status: "error",
+      error: "You cannot unblock yourself.",
+    });
+  }
+
+  BlockList.findOne(
+    { userId: userId, blockUserId: unblockUserId },
+    (err, unblockUser) => {
+      if (err || !unblockUser) {
+        return res.json({
+          data: null,
+          status: "error",
+          error: "You have already unblocked this user.",
+        });
+      }
+
+      User.findOne({ _id: userId }, (err, user) => {
+        if (err || !user) {
+          return res.json({
+            data: null,
+            status: "error",
+            error: "User not found in database.",
+          });
+        }
+
+        User.findOne({ _id: unblockUserId }, (err, unblockUser) => {
+          if (err) {
+            return res.json({
+              data: null,
+              status: "error",
+              error: "Unable to find unblock user.",
+            });
+          }
+
+          BlockList.deleteOne(
+            { blockUserId: unblockUserId, userId: userId },
+            (err, unblockUser) => {
+              if (err || !unblockUser) {
+                return res.json({
+                  data: null,
+                  status: "error",
+                  error: "Unable to find unblock user.",
+                });
+              }
+            }
+          );
+
+          return res.json({
+            data: `You have successfully unblocked this user.`,
+            status: "success",
+            error: null,
+          });
+        });
+      });
+    }
+  );
+};
+
 
 exports.unFollowUser = (req, res) => {
   const userId = req.profile._id;
